add user ownership for tasks

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -13,12 +14,18 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ) {}
 
-  getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto);
+  getTasks(
+    filterDto: GetTasksFilterDto,
+    user: User,
+  ): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user);
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const foundTask = await this.taskRepository.findOne(id);
+  async getTaskById(
+    id: number,
+    user: User,
+  ): Promise<Task> {
+    const foundTask = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 
     if (!foundTask) {
       throw new NotFoundException(`Task with ID "${id}" not found`);
@@ -27,8 +34,11 @@ export class TasksService {
     return foundTask;
   }
 
-  createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto);
+  createTask(
+    createTaskDto: CreateTaskDto,
+    user: User,
+  ): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user);
   }
   // getTaskById(id: string): Task {
   //   const foundTask = this.tasks.find(task => task.id === id);
@@ -40,17 +50,24 @@ export class TasksService {
   //   return foundTask;
   // }
   //
-  async deleteTaskById(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTaskById(
+    id: number,
+    user: User,
+  ): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
 
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID "${id}" not found`);
     }
   }
 
-  async updateTaskStatus(id: number, updateTaskStatusDto: UpdateTaskStatusDto): Promise<Task> {
+  async updateTaskStatus(
+    id: number,
+    updateTaskStatusDto: UpdateTaskStatusDto,
+    user: User,
+  ): Promise<Task> {
     const { status } = updateTaskStatusDto;
-    const task = await this.getTaskById(id);
+    const task = await this.getTaskById(id, user);
     Object.assign(task, { status });
     await task.save();
